requirements.txt
anthropic==0.52.2
mcp==1.9.3
playsound==1.2.2
python-dotenv==1.1.0
